requirements.txt
asyncpg>=0.27.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.22.0
